Moving color manager and tooltip builder to the service strategy.
